+++ frontend/techpick/src/components/dragOverlay.css.ts
@@ -33,18 +33,6 @@ export const stackedOverlay = style({
   position: 'relative',
 });
 
-export const stackedItem = style({
-  position: 'absolute',
-  width: '100%',
-  height: '100%',
-  background: 'white',
-  borderRadius: '4px',
-  border: '1px solid #ccc',
-  pointerEvents: 'none',
-  top: 0,
-  left: 0,
-});
-
 export const dragCount = style({
   position: 'absolute',
   top: '-8px',
